requirements.txt
gdown
scikit-learn==1.4.1.post1
numpy==1.26.4
pandas==2.2.1
matplotlib==3.8.3
pillow==10.2.0
transformers==4.39.3
--extra-index-url https://download.pytorch.org/whl/cu121
torch
torchvision
torchaudio